aggregate stale test into single group
