+++ test/use-swr-error.test.tsx
@@ -288,84 +288,6 @@ describe('useSWR - error', () => {
     screen.getByText('error: 1')
   })
 
-  it.skip('should not trigger the onLoadingSlow and onSuccess event after component unmount', async () => {
-    const key = createKey()
-    let loadingSlow = null,
-      success = null
-    function Page() {
-      const { data } = useSWR(key, () => createResponse('SWR'), {
-        onLoadingSlow: loadingKey => {
-          loadingSlow = loadingKey
-        },
-        onSuccess: (_, successKey) => {
-          success = successKey
-        },
-        loadingTimeout: 100
-      })
-      return <div>hello, {data}</div>
-    }
-
-    function App() {
-      const [on, toggle] = useState(true)
-      return (
-        <div id="app" onClick={() => toggle(s => !s)}>
-          {on && <Page />}
-        </div>
-      )
-    }
-
-    renderWithConfig(<App />)
-    screen.getByText('hello,')
-    expect(loadingSlow).toEqual(null)
-    expect(success).toEqual(null)
-
-    fireEvent.click(screen.getByText('hello,'))
-    await act(() => sleep(200))
-    expect(success).toEqual(null)
-    expect(loadingSlow).toEqual(null)
-  })
-
-  it.skip('should not trigger the onError and onErrorRetry event after component unmount', async () => {
-    const key = createKey()
-    let retry = null,
-      failed = null
-    function Page() {
-      const { data } = useSWR(key, () => createResponse(new Error('error!')), {
-        onError: (_, errorKey) => {
-          failed = errorKey
-        },
-        onErrorRetry: (_, errorKey) => {
-          retry = errorKey
-        },
-        dedupingInterval: 0
-      })
-      return (
-        <div>
-          <>hello, {data}</>
-        </div>
-      )
-    }
-
-    function App() {
-      const [on, toggle] = useState(true)
-      return (
-        <div id="app" onClick={() => toggle(s => !s)}>
-          {on && <Page />}
-        </div>
-      )
-    }
-
-    renderWithConfig(<App />)
-    screen.getByText('hello,')
-    expect(retry).toEqual(null)
-    expect(failed).toEqual(null)
-
-    fireEvent.click(screen.getByText('hello,'))
-    await act(() => sleep(200))
-    expect(retry).toEqual(null)
-    expect(failed).toEqual(null)
-  })
-
   it('should not trigger error retries if errorRetryCount is set to 0', async () => {
     const key = createKey()
     let count = 0
